requirements.txt
# Python dependencies for resume-ranker
pypdf>=3.0.0
python-docx>=0.8.11
sentence-transformers>=2.2.0
numpy>=1.21.0
pandas>=1.3.0
streamlit>=1.28.0
groq>=0.4.0
python-dotenv>=1.0.0

